+++ Business/Services/ICandidatesService.cs
@@ -7,7 +7,7 @@ namespace Business.Services
     public interface ICandidatesService
     {
         Task<List<CandidateDTO>> GetCandidatesAsync();
-        Task<int> GetAverageAsync();
+        Task<double> GetAverageAsync();
         Task<double> GetEcartTypeAsync();
         //Task<List<CandidateDTO>> CalculCandidatesNotes();
     }
